[CARBONDATA-2252] Query performance slows down as the number of columns increases in like query with OR expression #2060
Problem: In case of OR condition with like query contains and ends with, the filter is getting pushed down to carbon layer because of which the query is slow as compared to spark applying the same filter on the results returned from carbon

Analysis: This is because in case of like query the execution is done by RowLevelFilterExecutorImpl which will compute the data row by row. As the number of columns will increase the computation time will increase thereby increasing the query time.

Fix: If there is any OR condition with like query, it is better to return back all the results to spark and let spark do the computation.